RW-4554 add render tests
