+++ ui/src/app/cohort-search/cohort-search.module.ts
@@ -16,12 +16,11 @@ import {ListSearchComponent} from './list-search/list-search.component';
 import {ModalComponent} from './modal/modal.component';
 import {ModifierPageComponent} from './modifier-page/modifier-page.component';
 import {NodeInfoComponent} from './node-info/node-info.component';
-import {CriteriaTreeComponent} from './node/node.component';
 import {OptionInfoComponent} from './option-info/option-info.component';
 import {OverviewComponent} from './overview/overview.component';
 import {SearchGroupListComponent} from './search-group-list/search-group-list.component';
 import {SelectionInfoComponent} from './selection-info/selection-info.component';
-import {TreeComponent} from './tree/tree.component';
+import {CriteriaTreeComponent} from './tree/tree.component';
 
 import {BreadcrumbType} from 'app/utils/navigation';
 
@@ -69,7 +68,6 @@ const routes: Routes = [{
     SearchGroupListComponent,
     ListSearchComponent,
     SelectionInfoComponent,
-    TreeComponent,
     SafeHtmlPipe,
   ],
   providers: [TitleCasePipe]
