+++ ui/src/app/cohort-search/modal/modal.component.spec.ts
@@ -7,7 +7,7 @@ import {DemographicsComponent} from 'app/cohort-search/demographics/demographics
 import {ListSearchComponent} from 'app/cohort-search/list-search/list-search.component';
 import {ModifierPageComponent} from 'app/cohort-search/modifier-page/modifier-page.component';
 import {NodeInfoComponent} from 'app/cohort-search/node-info/node-info.component';
-import {NodeComponent} from 'app/cohort-search/node/node.component';
+import {NodeComponent} from 'app/cohort-search/tree-node/tree-node.component';
 import {OptionInfoComponent} from 'app/cohort-search/option-info/option-info.component';
 import {SafeHtmlPipe} from 'app/cohort-search/safe-html.pipe';
 import {SearchBarComponent} from 'app/cohort-search/search-bar/search-bar.component';
