--- ui/src/app/cohort-search/node/node.component.spec.ts
@@ -1,70 +1,0 @@
-import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-import {ReactiveFormsModule} from '@angular/forms';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {ClarityModule} from '@clr/angular';
-import {NodeInfoComponent} from 'app/cohort-search/node-info/node-info.component';
-import {SafeHtmlPipe} from 'app/cohort-search/safe-html.pipe';
-import {wizardStore} from 'app/cohort-search/search-state.service';
-import {registerApiClient} from 'app/services/swagger-fetch-clients';
-import {currentWorkspaceStore} from 'app/utils/navigation';
-import {CohortBuilderApi} from 'generated/fetch';
-import {NgxPopperModule} from 'ngx-popper';
-import {CohortBuilderServiceStub} from 'testing/stubs/cohort-builder-service-stub';
-import {workspaceDataStub} from 'testing/stubs/workspaces-api-stub';
-import {NodeComponent} from './node.component';
-
-describe('NodeComponent', () => {
-  let component: NodeComponent;
-  let fixture: ComponentFixture<NodeComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [
-        NodeInfoComponent,
-        NodeComponent,
-        SafeHtmlPipe,
-      ],
-      imports: [
-        BrowserAnimationsModule,
-        ClarityModule,
-        NgxPopperModule,
-        ReactiveFormsModule,
-      ],
-      providers: [],
-    })
-      .compileComponents();
-    currentWorkspaceStore.next({
-      ...workspaceDataStub,
-      cdrVersionId: '1',
-    });
-    wizardStore.next({});
-  }));
-
-  beforeEach(() => {
-    registerApiClient(CohortBuilderApi, new CohortBuilderServiceStub());
-    fixture = TestBed.createComponent(NodeComponent);
-    component = fixture.componentInstance;
-    component.node = {
-      code: '',
-      conceptId: 903133,
-      count: 0,
-      domainId: 'Measurement',
-      group: false,
-      hasAttributes: true,
-      id: 316305,
-      name: 'Height Detail',
-      parentId: 0,
-      predefinedAttributes: null,
-      selectable: true,
-      subtype: 'HEIGHT',
-      type: 'PM'
-    };
-    component.wizard = {};
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-
-});
