--- ui/src/app/cohort-search/tree/tree.component.spec.ts
@@ -1,75 +1,0 @@
-import {async, ComponentFixture, TestBed} from '@angular/core/testing';
-import {FormsModule, ReactiveFormsModule} from '@angular/forms';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {ClarityModule} from '@clr/angular';
-import {NodeInfoComponent} from 'app/cohort-search/node-info/node-info.component';
-import {NodeComponent} from 'app/cohort-search/node/node.component';
-import {OptionInfoComponent} from 'app/cohort-search/option-info/option-info.component';
-import {SafeHtmlPipe} from 'app/cohort-search/safe-html.pipe';
-import {SearchBarComponent} from 'app/cohort-search/search-bar/search-bar.component';
-import {registerApiClient} from 'app/services/swagger-fetch-clients';
-import {currentWorkspaceStore} from 'app/utils/navigation';
-import {CohortBuilderApi} from 'generated/fetch';
-import {NgxPopperModule} from 'ngx-popper';
-import {CohortBuilderServiceStub} from 'testing/stubs/cohort-builder-service-stub';
-import {workspaceDataStub} from 'testing/stubs/workspaces-api-stub';
-import {TreeComponent} from './tree.component';
-
-describe('TreeComponent', () => {
-  let component: TreeComponent;
-  let fixture: ComponentFixture<TreeComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [
-        NodeComponent,
-        NodeInfoComponent,
-        OptionInfoComponent,
-        SearchBarComponent,
-        TreeComponent,
-        SafeHtmlPipe,
-      ],
-      imports: [
-        BrowserAnimationsModule,
-        ClarityModule,
-        FormsModule,
-        NgxPopperModule,
-        ReactiveFormsModule,
-      ],
-      providers: [],
-    })
-      .compileComponents();
-    currentWorkspaceStore.next({
-      ...workspaceDataStub,
-      cdrVersionId: '1',
-    });
-  }));
-
-  beforeEach(() => {
-    registerApiClient(CohortBuilderApi, new CohortBuilderServiceStub());
-    fixture = TestBed.createComponent(TreeComponent);
-    component = fixture.componentInstance;
-    component.node = {
-      code: '',
-      conceptId: 903133,
-      count: 0,
-      domainId: 'Measurement',
-      group: false,
-      hasAttributes: true,
-      id: 316305,
-      name: 'Height Detail',
-      parentId: 0,
-      predefinedAttributes: null,
-      selectable: true,
-      subtype: 'HEIGHT',
-      type: 'PM'
-    };
-    component.wizard = {};
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-
-});
