+++ ui/src/app/cohort-search/tree/tree.component.spec.tsx
@@ -1,0 +1,24 @@
+import {shallow} from 'enzyme';
+import * as React from 'react';
+
+import {registerApiClient} from 'app/services/swagger-fetch-clients';
+import {currentWorkspaceStore} from 'app/utils/navigation';
+import {CohortBuilderApi} from 'generated/fetch';
+import {CohortBuilderServiceStub} from 'testing/stubs/cohort-builder-service-stub';
+import {workspaceDataStub} from 'testing/stubs/workspaces-api-stub';
+import {CriteriaTree} from './tree.component';
+
+describe('CriteriaTree', () => {
+  beforeEach(() => {
+    registerApiClient(CohortBuilderApi, new CohortBuilderServiceStub());
+    currentWorkspaceStore.next({
+      ...workspaceDataStub,
+      cdrVersionId: '1',
+    });
+  });
+  it('should create', () => {
+    const wrapper = shallow(<CriteriaTree/>)
+    expect(wrapper).toBeTruthy();
+  });
+
+});
